requirements.txt
streamlit
supabase
pandas
python-dotenv
psycopg2-binary
gspread
oauth2client
